commit parcial

+++ wallet/internal/adapters/output/database/extract/iextractdatabase.go
@@ -3,5 +3,5 @@ package database
 import "br.com.cleiton/wallet/internal/domain/entities"
 
 type IExtractDatabase interface {
-	GetExtract(walletId string, documentNumber string, page int64, size int64) ([]entities.Extract, error)
+	GetExtract(walletId string, documentNumber string, page int, size int) ([]entities.Extract, error)
 }
